requirements.txt
gradio
azure-cognitiveservices-vision-computervision
azure-ai-vision-imageanalysis
openai
azure-cognitiveservices-speech
python-dotenv
Pillow
requests